requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-user-agents==0.4.0
djangorestframework==3.12.4
Pillow==8.2.0
pkg-resources==0.0.0
promise==2.3
psycopg2-binary==2.8.6
python-memcached==1.59
pytz==2021.1
PyYAML==5.4.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.0
ua-parser==0.10.0
user-agents==2.2.0

